Finish SL-2603_CC_Feedback_rating_message

+++ frontend/views/client-chat-qa/view.php
@@ -5,6 +5,7 @@ use frontend\helpers\ChatHelper;
 use frontend\helpers\JsonHelper;
 use frontend\helpers\OutHelper;
 use sales\model\clientChat\entity\ClientChat;
+use sales\model\clientChatFeedback\entity\ClientChatFeedback;
 use sales\model\clientChatMessage\entity\ClientChatMessage;
 use sales\model\clientChatMessage\entity\search\ClientChatMessageSearch;
 use sales\model\clientChatNote\entity\ClientChatNote;
@@ -26,6 +27,7 @@ use yii\widgets\Pjax;
 /* @var yii\data\ActiveDataProvider $dataProviderNotes */
 /* @var yii\data\ActiveDataProvider $dataProviderRequest */
 /* @var ClientChatVisitorData|null $clientChatVisitorData */
+/* @var yii\data\ActiveDataProvider|null $dataProviderFeedback */
 
 $this->title = $model->cch_id;
 $this->params['breadcrumbs'][] = ['label' => 'Client Chats', 'url' => ['index']];
@@ -273,6 +275,9 @@ $this->params['breadcrumbs'][] = $this->title;
                             $view = '';
                             if (array_key_exists('attachments', $model->ccm_body)) {
                                 foreach ($model->ccm_body['attachments'] as $attachment) {
+                                    if (!is_array($attachment)) {
+                                        continue;
+                                    }
                                     if (array_key_exists('title_link', $attachment) && array_key_exists('title', $attachment)) {
                                         $titleLink = explode('.', $attachment['title_link']);
                                         $title = '[' . StringHelper::truncate($attachment['title'], 20) . '].' . end($titleLink);
@@ -311,6 +316,24 @@ $this->params['breadcrumbs'][] = $this->title;
                     ],
                     'ccn_user_id:userName',
                     'ccn_created_dt:byUserDateTime',
+                ],
+            ]);  ?>
+        </div>
+
+        <div class="col-md-6">
+            <h5>Chat feedback</h5>
+            <?php echo GridView::widget([
+                'dataProvider' => $dataProviderFeedback,
+                'columns' => [
+                    [
+                        'attribute' => 'ccf_message',
+                        'value' => static function (ClientChatFeedback $feedback) {
+                            return Html::encode($feedback->ccf_message);
+                        },
+                    ],
+                    'ccf_rating',
+                    'ccf_user_id:userName',
+                    'ccf_created_dt:byUserDateTime',
                 ],
             ]);  ?>
         </div>
